# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.PremiumActivateRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  dto.PremiumStatusResponse:
    properties:
      active:
        type: boolean
      days_remaining:
        description: 0=普通会员, -1=永久, >0=剩余天数
        type: integer
      start_time:
        type: string
    type: object
...
  dto.SettingGroupResponse:
    properties:
      group:
        type: string
      settings:
        items:
          $ref: '#/definitions/dto.SettingItemResponse'
        type: array
    type: object
  dto.SettingItemResponse:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  dto.UpdateSettingGroupRequest:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    required:
    - settings
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FlecBlog
  version: v1
paths:
...
  /admin/settings/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定分组的所有配置项
      parameters:
      - description: 配置分组
        enum:
        - basic
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SettingGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取配置分组
      tags:
      - 配置管理
    patch:
      consumes:
      - application/json
      description: 批量更新指定分组的配置项，限超级管理员
      parameters:
      - description: 配置分组
        enum:
        - basic
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      - description: 配置更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSettingGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新配置分组
      tags:
      - 配置管理
  /admin/settings/ai/mcp-secret/reset:
    put:
      consumes:
      - application/json
      description: 重新生成并返回 MCP Secret，限超级管理员
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 重置 MCP Secret
      tags:
      - 配置管理
...
  /api/v1/admin/premium/activate:
    post:
      consumes:
      - application/json
      description: 用户输入激活码，Server 调 Panel 验证后累加会员时间
      parameters:
      - description: 激活请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PremiumActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PremiumStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 激活会员
      tags:
      - 会员
  /api/v1/admin/premium/status:
    get:
      description: 获取当前站点的会员状态信息（包含完整字段，仅管理后台用）
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PremiumStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询会员状态
      tags:
      - 会员
...
  /settings/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定分组的公开配置项，根据配置项可见性规则筛选
      parameters:
      - description: 配置分组
        enum:
        - basic
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取配置分组
      tags:
      - 配置
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"